+++ src/core/export.jl
@@ -47,3 +47,10 @@ for status_code_enum in [TerminationStatusCode, ResultStatusCode]
     end
 end
 
+# from InfrastructureModels
+export ids, ref, var, con, sol, nw_ids, nws, optimize_model!
+
+# for backward compat with previous versions, drop on PowerModels v0.16
+import InfrastructureModels: sol_component_fixed, sol_component_value, sol_component_value_edge, build_solution, build_result, report_duals, ismultinetwork
+export sol_component_fixed, sol_component_value, sol_component_value_edge, build_solution, build_result, report_duals, ismultinetwork
+