+++ src/form/acp.jl
@@ -81,7 +81,7 @@ function constraint_power_balance(pm::AbstractACPModel, n::Int, i::Int, bus_arcs
         + sum(bs for (i,bs) in bus_bs)*vm^2
     )
 
-    if report_duals(pm)
+    if _IM.report_duals(pm)
         sol(pm, n, :bus, i)[:lam_kcl_r] = cstr_p
         sol(pm, n, :bus, i)[:lam_kcl_i] = cstr_q
     end
@@ -149,7 +149,7 @@ function constraint_power_balance_ls(pm::AbstractACPModel, n::Int, i::Int, bus_a
         )
     end
 
-    if report_duals(pm)
+    if _IM.report_duals(pm)
         sol(pm, n, :bus, i)[:lam_kcl_r] = cstr_p
         sol(pm, n, :bus, i)[:lam_kcl_i] = cstr_q
     end