+++ src/io/matpower.jl
@@ -125,7 +125,7 @@ const _mp_switch_columns = [
 
 ""
 function _parse_matpower_string(data_string::String)
-    matlab_data, func_name, colnames = InfrastructureModels.parse_matlab_string(data_string, extended=true)
+    matlab_data, func_name, colnames = _IM.parse_matlab_string(data_string, extended=true)
 
     case = Dict{String,Any}()
 
@@ -155,8 +155,8 @@ function _parse_matpower_string(data_string::String)
     if haskey(matlab_data, "mpc.bus")
         buses = []
         for bus_row in matlab_data["mpc.bus"]
-            bus_data = InfrastructureModels.row_to_typed_dict(bus_row, _mp_bus_columns)
-            bus_data["index"] = InfrastructureModels.check_type(Int, bus_row[1])
+            bus_data = _IM.row_to_typed_dict(bus_row, _mp_bus_columns)
+            bus_data["index"] = _IM.check_type(Int, bus_row[1])
             bus_data["source_id"] = ["bus", bus_data["index"]]
             push!(buses, bus_data)
         end
@@ -168,7 +168,7 @@ function _parse_matpower_string(data_string::String)
     if haskey(matlab_data, "mpc.gen")
         gens = []
         for (i, gen_row) in enumerate(matlab_data["mpc.gen"])
-            gen_data = InfrastructureModels.row_to_typed_dict(gen_row, _mp_gen_columns)
+            gen_data = _IM.row_to_typed_dict(gen_row, _mp_gen_columns)
             gen_data["index"] = i
             gen_data["source_id"] = ["gen", i]
             push!(gens, gen_data)
@@ -181,7 +181,7 @@ function _parse_matpower_string(data_string::String)
     if haskey(matlab_data, "mpc.branch")
         branches = []
         for (i, branch_row) in enumerate(matlab_data["mpc.branch"])
-            branch_data = InfrastructureModels.row_to_typed_dict(branch_row, _mp_branch_columns)
+            branch_data = _IM.row_to_typed_dict(branch_row, _mp_branch_columns)
             branch_data["index"] = i
             branch_data["source_id"] = ["branch", i]
             push!(branches, branch_data)
@@ -194,7 +194,7 @@ function _parse_matpower_string(data_string::String)
     if haskey(matlab_data, "mpc.dcline")
         dclines = []
         for (i, dcline_row) in enumerate(matlab_data["mpc.dcline"])
-            dcline_data = InfrastructureModels.row_to_typed_dict(dcline_row, _mp_dcline_columns)
+            dcline_data = _IM.row_to_typed_dict(dcline_row, _mp_dcline_columns)
             dcline_data["index"] = i
             dcline_data["source_id"] = ["dcline", i]
             push!(dclines, dcline_data)
@@ -205,7 +205,7 @@ function _parse_matpower_string(data_string::String)
     if haskey(matlab_data, "mpc.storage")
         storage = []
         for (i, storage_row) in enumerate(matlab_data["mpc.storage"])
-            storage_data = InfrastructureModels.row_to_typed_dict(storage_row, _mp_storage_columns)
+            storage_data = _IM.row_to_typed_dict(storage_row, _mp_storage_columns)
             storage_data["index"] = i
             storage_data["source_id"] = ["storage", i]
             push!(storage, storage_data)
@@ -216,7 +216,7 @@ function _parse_matpower_string(data_string::String)
     if haskey(matlab_data, "mpc.switch")
         switch = []
         for (i, switch_row) in enumerate(matlab_data["mpc.switch"])
-            switch_data = InfrastructureModels.row_to_typed_dict(switch_row, _mp_switch_columns)
+            switch_data = _IM.row_to_typed_dict(switch_row, _mp_switch_columns)
             switch_data["index"] = i
             switch_data["source_id"] = ["switch", i]
             push!(switch, switch_data)
@@ -227,7 +227,7 @@ function _parse_matpower_string(data_string::String)
     if haskey(matlab_data, "mpc.bus_name")
         bus_names = []
         for (i, bus_name_row) in enumerate(matlab_data["mpc.bus_name"])
-            bus_name_data = InfrastructureModels.row_to_typed_dict(bus_name_row, _mp_bus_name_columns)
+            bus_name_data = _IM.row_to_typed_dict(bus_name_row, _mp_bus_name_columns)
             bus_name_data["index"] = i
             bus_name_data["source_id"] = ["bus_name", i]
             push!(bus_names, bus_name_data)
@@ -281,7 +281,7 @@ function _parse_matpower_string(data_string::String)
                 end
                 tbl = []
                 for (i, row) in enumerate(matlab_data[k])
-                    row_data = InfrastructureModels.row_to_dict(row, column_names)
+                    row_data = _IM.row_to_dict(row, column_names)
                     row_data["index"] = i
                     row_data["source_id"] = [case_name, i]
                     push!(tbl, row_data)
@@ -309,16 +309,16 @@ function _mp_cost_data(cost_row)
        nr_parameters = ncost
     end
     cost_data = Dict(
-        "model" => InfrastructureModels.check_type(Int, cost_row[1]),
-        "startup" => InfrastructureModels.check_type(Float64, cost_row[2]),
-        "shutdown" => InfrastructureModels.check_type(Float64, cost_row[3]),
-        "ncost" => InfrastructureModels.check_type(Int, cost_row[4]),
-        "cost" => [InfrastructureModels.check_type(Float64, x) for x in cost_row[5:5+nr_parameters-1]]
+        "model" => _IM.check_type(Int, cost_row[1]),
+        "startup" => _IM.check_type(Float64, cost_row[2]),
+        "shutdown" => _IM.check_type(Float64, cost_row[3]),
+        "ncost" => _IM.check_type(Int, cost_row[4]),
+        "cost" => [_IM.check_type(Float64, x) for x in cost_row[5:5+nr_parameters-1]]
     )
 
     #=
     # skip this literal interpretation, as its hard to invert
-    cost_values = [InfrastructureModels.check_type(Float64, x) for x in cost_row[5:length(cost_row)]]
+    cost_values = [_IM.check_type(Float64, x) for x in cost_row[5:length(cost_row)]]
     if cost_data["model"] == 1:
         if length(cost_values)%2 != 0
             Memento.error(_LOGGER, "incorrect matpower file, odd number of pwl cost function values")
@@ -380,7 +380,7 @@ function _matpower_to_powermodels!(mp_data::Dict{String,<:Any})
     _split_loads_shunts!(pm_data)
 
     # use once available
-    InfrastructureModels.arrays_to_dicts!(pm_data)
+    _IM.arrays_to_dicts!(pm_data)
 
     for optional in ["dcline", "load", "shunt", "storage", "switch"]
         if length(pm_data[optional]) == 0