+++ src/prob/opb.jl
@@ -22,7 +22,7 @@ end
 
 
 function ref_add_connected_components!(pm::AbstractPowerModel)
-    if InfrastructureModels.ismultinetwork(pm.data)
+    if _IM.ismultinetwork(pm.data)
         nws_data = pm.data["nw"]
     else
         nws_data = Dict("0" => pm.data)