+++ CHANGELOG.md
@@ -2,6 +2,7 @@ PowerModels.jl Change Log
 =========================
 
 ### Staged
+- Updated to utilize generic base functions from InfrastructureModels
 - Fixed the type of derived "ncost" values in `correct_cost_functions!`
 
 ### v0.15.3